nixl_ep tests: validate new flags, replace hard-coded 8, update copyright
